#160 - rename monaco template providers to kebab-case
