+++ src/renderer/lib/monaco/template-variable-completion-items-provider.ts
