+++ src/renderer/lib/monaco/template-variable-hover-provider.ts
