+++ src/renderer/lib/monaco/template-variable-semantic-tokens-provider.ts
